+++ bundler/spec/install/gems/mirror_probe_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe "fetching dependencies with a not available mirror", realworld: true do
+RSpec.describe "fetching dependencies with a not available mirror" do
   let(:mirror) { @mirror_uri }
   let(:original) { @server_uri }
   let(:server_port) { @server_port }
@@ -108,8 +108,8 @@ RSpec.describe "fetching dependencies with a not available mirror", realworld: t
     @server_port = find_unused_port
     @server_uri = "http://#{host}:#{@server_port}"
 
-    require_relative "../support/artifice/endpoint"
-    require_relative "../support/silent_logger"
+    require_relative "../../support/artifice/endpoint"
+    require_relative "../../support/silent_logger"
 
     require "rackup/server"
